+++ src/js/components/award/financialAssistance/CFDASection.jsx
@@ -5,42 +5,15 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import AwardSection from '../shared/AwardSection';
 import AwardSectionHeader from '../shared/AwardSectionHeader';
 import { CFDASectionInfo } from "../shared/InfoTooltipContent";
-import ExpandableAwardSection from '../shared/ExpandableAwardSection';
+import SingleCFDA from './SingleCFDA';
 
 const propTypes = {
-    data: PropTypes.shape({
-        cfda_number: PropTypes.string,
-        cfda_title: PropTypes.string,
-        applicant_eligibility: PropTypes.string,
-        beneficiary_eligibility: PropTypes.string,
-        federal_agency: PropTypes.string,
-        objectives: PropTypes.string,
-        obligations: PropTypes.string,
-        popular_name: PropTypes.string,
-        website: PropTypes.string,
-        sam_website: PropTypes.string
-    })
+
 };
 
-const CFDASection = ({ data }) => {
-    const {
-        samWebsite,
-        cfdaWebsite,
-        cfdaFederalAgency,
-        cfdaNumber,
-        cfdaTitle,
-        applicantEligibility,
-        beneficiaryEligibility,
-        cfdaObjectives
-    } = data;
-    const expandableContent = (
-        <React.Fragment>
-            <h5>Applicant Eligibility</h5>
-            <ExpandableAwardSection type="secondary" content={applicantEligibility} />
-            <h5>Beneficiary Eligibility</h5>
-            <ExpandableAwardSection type="secondary" content={beneficiaryEligibility} />
-        </React.Fragment>
-    );
+const CFDASection = ({
+
+}) => {
     return (
         <AwardSection type="column" className="cfda-section award-viz">
             <AwardSectionHeader
@@ -50,17 +23,7 @@ const CFDASection = ({ data }) => {
                 left={false}
                 tooltipWide />
             <div className="award__col__content">
-                <ExpandableAwardSection content={expandableContent}>
-                    <h4>{`${cfdaNumber}: ${cfdaTitle.toUpperCase()}`}</h4>
-                    <h5>Objectives</h5>
-                    <ExpandableAwardSection type="secondary" content={cfdaObjectives} />
-                    <h5>Administrative Agency</h5>
-                    <p>{cfdaFederalAgency}</p>
-                    <h5>Website</h5>
-                    <a href={cfdaWebsite}>{cfdaWebsite}</a>
-                    <h5>SAM.gov Page</h5>
-                    <a href={samWebsite}>{samWebsite}</a>
-                </ExpandableAwardSection>
+              hi
             </div>
         </AwardSection>
     );
